versionMatcher .cpp and .hpp files were exploded into smaller independent class files. Constructors and operators were refactored to improve code clearity.
