+++ src/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/CMakeLists.txt
@@ -6,6 +6,7 @@ enable_testing()
 
 file(GLOB DATABASE_FEED_SRC
     "*.cpp"
+    "version_matcher/*.cpp"
     )
 
 # Add fpic flag
