+++ src/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/version_matcher/versionObjectPEP440.cpp
@@ -1,7 +1,7 @@
 /*
  * Wazuh Vulnerability scanner - Database Feed Manager
  * Copyright (C) 2015, Wazuh Inc.
- * October 31, 2023.
+ * November 3, 2023.
  *
  * This program is free software; you can redistribute it
  * and/or modify it under the terms of the GNU General Public
@@ -9,15 +9,8 @@
  * Foundation.
  */
 
-#include "versionMatcher.hpp"
-
-std::regex VersionObjectCalVer::parserRegex(R"((\d{2}|\d{4})(\.\d{1,2})?(\.\d{1,2})?(\.\d+)?)");
+#include "versionObjectPEP440.hpp"
 
 std::regex VersionObjectPEP440::parserRegex(
     R"(^(?:([1-9][0-9]*)!)?((?:0|[1-9][0-9]*)(?:\.(?:0|[1-9][0-9]*))*)(?:(a|b|rc)(0|[1-9][0-9]*))?(?:\.post(0|[1-9][0-9]*))?(?:\.dev(0|[1-9][0-9]*))?$)");
 std::regex VersionObjectPEP440::parserVersionStrRegex(R"(\.)");
-
-std::regex VersionObjectMajorMinor::parserRegex(R"(^(\d+)[.\-](\d+)$)");
-
-std::regex VersionObjectSemVer::parserRegex(
-    R"(^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$)");
