+++ src/wazuh_modules/vulnerability_scanner/tests/databaseFeedManager/CMakeLists.txt
@@ -9,7 +9,8 @@ file(GLOB DATABASE_FEED_MANAGER_UNIT_TEST_SRC
     "*.cpp")
 
 file(GLOB DATABASEFEED_MANAGER_SRC
-    "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/*.cpp")
+    "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/*.cpp"
+    "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/version_matcher/*.cpp")
 
 add_definitions(-DWAZUH_UNIT_TESTING)
 
@@ -23,7 +24,9 @@ add_dependencies(database_feed_manager_unit_test compile_schemas) #Add a depende
 
 target_compile_definitions(database_feed_manager_unit_test PUBLIC FLATBUFFER_SCHEMAS_DIR="${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/schemas")
 
-target_include_directories(database_feed_manager_unit_test PRIVATE "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager")
+target_include_directories(database_feed_manager_unit_test 
+    PRIVATE "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager"
+    PRIVATE "${SRC_FOLDER}/wazuh_modules/vulnerability_scanner/src/databaseFeedManager/version_matcher")
 
 target_link_libraries(database_feed_manager_unit_test
     debug gtestd
