+++ src/wazuh_modules/vulnerability_scanner/tests/databaseFeedManager/versionMatcher_test.cpp
@@ -8,7 +8,7 @@
  * License (version 2) as published by the FSF - Free Software
  * Foundation.
  */
-#include "versionMatcher_test.h"
+#include "versionMatcher_test.hpp"
 
 using ::testing::_;
 using ::testing::Return;
