+++ src/wazuh_modules/vulnerability_scanner/tests/databaseFeedManager/versionMatcher_test.hpp
@@ -8,9 +8,9 @@
  * License (version 2) as published by the FSF - Free Software
  * Foundation.
  */
-#ifndef _VERSION_MATCHER_TEST_H
-#define _VERSION_MATCHER_TEST_H
-#include "../../src/databaseFeedManager/versionMatcher.hpp"
+#ifndef _VERSION_MATCHER_TEST_HPP
+#define _VERSION_MATCHER_TEST_HPP
+#include "versionMatcher.hpp"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
@@ -45,4 +45,4 @@ protected:
     void TearDown() override;
 };
 
-#endif //_VERSION_MATCHER_TEST_H
+#endif //_VERSION_MATCHER_TEST_HPP
